Fix futex-related deadlocks.
1) Do not perform system calls, in this case clock_gettime, while
holding the futex ticketlock. It will make the thread yield while
holding the lock. If subsequently all ethreads are occupied by threads
trying to acquire the lock, these threads will spin forever, and the
lock-holding thread is never scheduled again to release the lock.
Alternatively, the ticketlock implementation can be changed to yield
after some time.  However, the time holding a ticketlock should be as
short as possible, so that performing syscalls or yielding in general
while holding the lock should be avoided where possible.

2) Both free and malloc calls can under certain circumstances result in
futex calls. Calling free and/or malloc as part of the futex
implementation while holding the lock will therefore lead to a deadlock
as the ticketlock-holding thread is trying to acquire the lock again. We
now pre-allocate memory for the required futex_q structure as part of
the lthread struct instead of dynamically allocating and freeing memory
as part of the futex implementation.