Run java with -XX:+UseMembar by default.
The JVM uses a "serialization page" in a similar way to the polling
page. The VM thread revokes write permissions from the serialization
page, which will cause segfaults in Java threads and allow them to get
to a safepoint. It's used to get a consistent view over the state of all
Java threads.

With -XX:+UseMembar the JVM uses actual memory barriers and does not use
the page fault mechanism to achieve the same. However, it is disabled by
default.  It is unlikely that the page fault mechanism has better
performance with SGX-LKL as every segfault will result in at least two
enclave exits (one asynchronous and one synchronous) and two enclave
entries (one EENTER and one ERESUME). We therefore enable it by default.