Prevent threads from reentering the enclave while the enclave is being destroyed/unmapped.
This might still not be completely thread-safe, as one thread can be in the
process of re-entering, while another thread is terminating the enclave
(potentially leading to a segfault during ENCLU). Since this is happening
during termination, this should be good enough though as the enclave will be
terminated in either case and the race condition should be rare.