Pass in errno value (negated) as return value for host system calls.
musl's __syscall function is expected to return the (negated) errno value as
return value. Therefore setting errno and setting the return value to -1 should
not be done as part of the host sytem call handling. This is done by the
syscall_ret function (within the enclave) instead. Calls to syscall (not
__syscall) will have set the errno value as expected.