Fix hangs due to debug print ticket lock.
There are two possible race conditions/deadlocks due to the debug ticket lock.

1) The lock-holding thread yields when doing the write system call to print to
stdout. Other threads trying to obtain the same lock will spin until the first
thread is scheduled again and will release the lock. However, if all physical
threads are occupied by enclave threads waiting for the lock, the lock-holding
thread will never get scheduled againd and can't release the lock.

2) If logging of threading-related events (via SGXLKL_TRACE_THREAD) is enabled,
the lthread scheduler logs an event before re-scheduling a thread which has
been waiting for a host syscall to return. This includes threads that are
currently holding the debug print ticket lock. The scheduler is waiting for the
same lock to be release and at the same time is holding off from scheduling the
lock-holding thread, causing a deadlock.

Threads should not yield (e.g. when doing a host system call) while holding a
ticket lock. Therefore, we remove the debug print ticket lock, and add spin
locks for writes to stdout and stderr instead outside of the enclave instead.