Check for enclave signing key when SGXLKL_HEAP is specified.
Previously SGXLKL_HEAP, used to set the enclave heap size, was silently
ignored when no signing key was provided via SGXLKL_KEY. We now print an
error and terminate when no signing key is provided. In cases where
SGXLKL_HEAP and SGXLKL_KEY are not specified, the heap size is
calculated based on the amount of available memory in the EPC. As part
of the build process the enclave is signed with a development key for
this default heap size.