Fix loading of libsgxlkl.so with multiple loadable segments.
With newer version of LKL liblkl.a/libsgxlkl.so has multiple loadable segments.
This causes issues in simulation mode when mapping these segments into memory.
Segments are expected to be mapped into adjacent regions of memory with fixed
offsets. Therefore the corresponding mappings are created using mmap with the
MAP_FIXED flag. With this commit we ensure that there cannot be any unrelated
existing mappings that would overlap with the mapping for loadable segments by
determining the total size of all loadable segments upfront and creating a
mapping of that size when we map the first loadable segment.