Add support for secondary disks.
Additional disks can be specified via SGXLKL_HDS which expects the following
format:

  SGXLKL_HDS=disk1path:disk1mntpoint:disk1ro,disk2path:disk2mntpoint:disk2ro,[...]

e.g.

  SGXLKL_HDS="./appdata.img:/mnt/data:0" ./build/sgx-lkl-run ../sgxlkl-root-disk.img /usr/bin/myapp /mnt/data/myinput

Multiple disk specifications are separated by a comma (","). Each specification
consists of three parts separated by a colon (":"). The first value is the path
to the disk image file on the host system. The second value specifies the
in-enclave mount point through which the disk will be accessible at runtime.
The third value specifies whether the disk should be mounted as read-only
(value "1") or as writeable (any other value).

As before, the root disk is specified as first argument to sgx-lkl-run as it is
mandatory.

/mnt is mounted as tmpfs to allow mounting of secondary file systems when the
root file system is mounted as read-only.