Revert to Alpine v3.8.2 in sgx-lkl-disk for now.
Alpine v3.9 and v.3.10 are based on kernel v4.19 while SGX-LKL is currently
based on 4.17. To prevent incompatibilities with applications built against a
newer kernel version, revert to the Alpine v3.8.2 (based on v4.14) root image
and package repositories when used with --alpine until SGX-LKL is updated to
v4.19+.

Fixes #51.