Add CI (#101)
Various changes to the CI scripts to make them work on a stock Ubuntu 18.04 image. Note that the $USER work-around is only temporary and is being fixed in Azure Pipelines.

Also, note that I changed the submodule URLs to use http consistently so that it can be cloned without private ssh key being available. This allows to run CI on external PR builds without allowing access to any secrets. If you want to use ssh locally for the submodules then you can still do that without changing the URLs, see https://stackoverflow.com/a/31185110.