Provide an in-enclave source of time (#162)
Prior to this commit, sgx-lkl didn't have an in-enclave source of time. When the
enclave needed to observe the passage of time, it would execute ocalls out to
the host in order to get it. This approach was slow and brittle. It also
resulted in the existence of a number of bugs related to various time related
system calls such as time and utime.


This change provides an in-enclave time source by adding a monotonic source of
nanosecond passage of time. Linux uses said time source as the underpinning for
other time related sources and bubbles up to application programmer level 
concerns such as `gettimeofday`, `clock_settime`, `clock_gettime`, `time`, `utime`, etc.