LTP: Fix test issue fcntl01,fcntl12 (#226)
* LTP:Fix test issue fcntl01

Test case is marked as failed because no Pass/Fail message
printed in (stdout and stderr) logs. Failure test messages
are already present in case of failed condition in the program.
Hence, added the PASS message at the end of the test.

* LTP: Fix issue in testcase fcntl12

Test case (testcases/kernel/syscalls/fcntl/fcntl12)
is failed/skipped because it exit improperly in middle of test.
this is due to single process environment. Test case is modified
to work in single process environment.