Remove some libc usage in src/enclave
This commit removes some of the malloc/calloc/free/strdup libc usage
in src/enclave. Not all usage as been removed yet. There are other
cases that I have left because when changed, the test suite segfaults
for non-obvious reasons that need to be looked into.

This commit addresses a small part of #151.