Compile OE with COMPILE_SYSTEM_EDL
Recently there has been work done in OE to convert ecalls and ocalls that
were previously builtin to core libraries and make them opt-in instead.
This change will mean that SGX-LKL will need to specifically import the
EDL that it uses from oecore.

I am turning off this new feature to prevent upcoming changes from
breaking SGX-LKL. Once the feature is fully merged in OE I will revert
this change and make the necessary changes in SGX-LKL to make use of the
feature.

Signed-off-by: Jordan Hand <jorhand@microsoft.com>