Use the LKL futex implementation in userspace.
We currently have two implementations of futexes, one provided by Linux
(which was not being built) and one in `src/enclave`.  The `src/enclave`
version was incomplete (for example, not implementing `FUTEX_WAKE_OP`)
and was under-tested.  After this change:

 - The `enclave_futex` implementation is used only for the LKL host ops.
 - The Linux futex is used for everything in userspace.

This should now give us parity with Linux for futex operations (and
ensure that we pick up new ones when they are added upstream, for
example the proposed operation to wait on multiple futexes).

This does not fix #209, but hopefully simplifies the fix because there
should now be only one way for a userspace thread to block.  It appears
as if the thread is blocking correctly in userspace and the scheduler is
receiving ticks, but is not waking up the blocking thread.

Fixes #180
Fixes #305