Move the LKL host layer to using OE malloc.
I tried this previously, but reverted it because of crashes.  The
crashes were caused by the fact that LKL allocate memory for the kernel
and for `lkl_vprintf` buffers.  The former is larger than the OE heap,
so fails, resulting in LKL refusing to start.  To avoid this, assume
that there is only going to be one allocation > 1MiB and that this is
the kernel heap.  Special case it and provide it via `enclave_mmap`,
forward everything else to `oe_malloc`.  Other structures should be safe
to allocate via `oe_malloc`.