Fix race condition in clone test
Prior to this change, it was possible for the final result of the addition
to end up as 101 instead of 100 which would subsequently cause the test
to hang as it would later wait on the value being 100.

The race condition was this:

Between when a thread read that the value of counter was less than 100 and when it
incremented the value, the value might have been changed in the other thread.