Fix bug in sem_up/sem_down implementation
Prior to this change, it is possible to lose wakeup events.
For example, 3 different things call sem_down resulting in a
sem count of 0.

On waking, only 1 would be awoken. The rest would continue to sleep.

This changes so that all are awoken and they can see which gets
the semaphore and then go back to sleep. Because all are awoken, the
other logic works.