Defer unmapping the thread stack.
When a thread is created with clone and detached, the last thing that
libc does before the thread exits is unmap the stack.  This works on
most Linux architectures because the last bit of assembly code can put
the thread exit code in a register and perform an exit system call.  It
does not work with LKL, because LKL does not maintain separate kernel
and userspace threads and because LKL passes arguments in memory (which,
in turn, means that the caller typically allocates them on the stack).

To work around this, we check on `munmap` calls whether the region being
unmapped contains the stack pointer.  If it does, then we defer
unmapping the stack until lthreads cleans up the thread (at which point,
nothing is using it and the unmap operation runs on a different thread's
stack).