Remove bad munmap03 LTP test patch
The patch is wrong. It hides a bug in munmap. On failure, munmap is
supposed to return -1. The patch as noted in its comment allows for
0 aka success to be returned and assumes because the errno was set
correctly, that everything is correct.

It isn't. The return value has to be -1 when failure occurs. Anyone
using this munmap interface would get surprising results and the test
patch hides that. munmap is currently broken.