Label all newly opened issues as "needs-triage" (#692)
* Label all newly opened issues as "needs-triage"

This commit adds a GitHub action to label all newly opened issues with the "needs-triage" label.
This is the simplest way forward with automating this. This uses a 3rd party GitHub action that is pinned
to specific SHA hash.

A larger project would be to replace this with a custom written GitHub action that runs on a schedule using
the GitHub actions cron facilities. That action would iterate through all issues in the repo that don't have a `p*` priority 
label and add the `needs-triage` label. 

This approach which is much simpler and doesn't require any real investment of time.

For this to work, if GitHub actions (specifically running 3rd party ones) aren't enabled for the lsds/sgx-lkl repo,
they will need to be turned on.

* Fix typo

* Remove blank line