Remove dead enclave futex code
enclave_futex is only used from posix-host.c. None of the code removed
in this PR is live.

Without the fallthroughs needed by dead code, I decided that an if/else
block would be better than a switch so I made some changes to structure
in enclave_futex after the changes.