Remove enclave_futex and narrow the enclave provided futex API (#708)
The prior version of enclave_futex was left over from when we had a
"complete" futex implementation in the enclave. At one point, it was
used as the futex implementation for everything flowing through musl.

That was changed, however, enclave_futex lived on and continued to
accept an operation to perform (e.g. FUTEX_WAKE). Not every operation
from the futex API was guaranteed to work.

To make matters worse, if the caller didn't check the result of
enclave_futex, it is possible they would miss that nothing was done
for unimplemented futex operations.

This commit removes enclave_futex and replaces it with 3 targeted
API calls enclave_futex_wake, enclave_futex_wait, and enclave_futex_timedwait
between the 3, we have complete coverage of all futex functionality currently
used to support internal enclave functionality.