Add a brute-force temporal-safety checking allocator. (#719)
Optionally enable it for lkl_sem and lthread structures.

This allocates each object on a separate page and uses mprotect on the
host to remove all permissions to access the object after it's
deallocated.  In production, this would quickly exhaust enclave memory,
but it can be useful when running tests.