chmod rootfs files when mounting
This fixes #236.

The problem was that we were setting "/" to be 700 for perms which
doesn't allow any user other than root to access "/tmp" and
other directories.

After some review and conversation with @prp, I believe that in general
the existing code in lkl_prepare_rootfs was there as a "just in case"
sanity check. It creates directories if they are missing and provides
some permissions (which differ from normal Linux installs). However,
those directories almost always existed.

This change will now chmod those directories (including "/") if they
exit. Additionally it updates the perms to be set to be 755 to match
a standard Linux environment.