Don't overwrite LTP tests when doing make clean
Prior to this change when doing `make clean` in the `ltp` directory,
it would resync the ltp git submodule. This is highly problematic
for any workflow where you would need to work with a different commit
of the ltp submodule or if you want to make changes in tree to test.

For example, if debugging a problem with an ltp test.

You modify the ltp test.
Run make clean and go to run the given test.

Prior to this change, all your modifications to the ltp test would be
overwritten.

As of this commit, the ltp submodule will only be sync'd if it hasn't
already been pulled down. This should keep existing workflows operating
while also making submodule checkout for ltp work the same as those for
sgx-lkl-musl etc.