Unify LTP Makefiles
Prior to this commit, there were two sets of almost identical "Makefiles"
for LTP.

batch.mk that was used for the Makefiles in ltp-batch* and Makefile that
is currently unused in CI, but can be used to run tests manually from the
LTP directory.

This commit changes to have both sets of "Makefiles" use batch.mk by
merging the few small differences in relatives paths into batch.mk and
checking for the files are known locations and setting up accordindly.