Fix performance collapse in network.
The code instructing the polling thread to poll had a race condition
that meant that sometimes the polling thread would poll only for reads
or writes (or neither, just for subsequent wake-ups).

With this, the iperf test in #777 runs to completion with an average of
around 3Gb/s.

Fixes #681

Related to #499