Don't rebuilt LTP LTP_TEST_MNT_IMG ever time a test is run
For the Makefile rule to correctly work and note that ltp_tst_mntfs.img has been created, it needs to not be in quotes. 
Otherwise it is looking for the existence of the file in quotes and causes the check to fail which results in root fs being rebuilt as well.