[Test] Add test for /proc/cpuinfo and sysinfo()
This test makes sure that the number of CPUs is the same number as
ethreads. The environment variable SGXLKL_ETHREADS is set to an
arbitrary value (in the current code, 4) and the same number is passed
as a command line argument to the binary ``cpuinfo``, which verifies
that the number of CPUs in the emulated system matches the number passed
on the command line.