Clean up the ticket lock implementation.
This reduces the amount of undefined behaviour in the ticket lock
considerably.  I don't remember what WG14 eventually decided on atomic
operations of different sizes at the same address, I believe this is
non-UB (it's just broken on some POWER CPUs), but it is at least safe on
x86.

At the same time, add some comments explaining what the code is trying
to do so that the next person who needs to understand it takes less time
than me.